remove es256k from oauth package
