+++ crypto/jwx/jwx_es256k.go
@@ -18,28 +18,10 @@
  *
  */
 
-package crypto
+package jwx
 
-import (
-	"crypto"
-	"github.com/lestrrat-go/jwx/v2/jwa"
-	"github.com/lestrrat-go/jwx/v2/jwt"
-	"github.com/nuts-foundation/nuts-node/crypto/test"
-	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
-	"testing"
-)
+import "github.com/lestrrat-go/jwx/v2/jwa"
 
-func TestES256k(t *testing.T) {
-	t.Run("test ES256K", func(t *testing.T) {
-		ecKey := test.GenerateECKey()
-		token := jwt.New()
-		signature, _ := jwt.Sign(token, jwt.WithKey(jwa.ES256K, ecKey))
-		parsedToken, err := ParseJWT(string(signature), func(_ string) (crypto.PublicKey, error) {
-			return ecKey.Public(), nil
-		})
-		require.NoError(t, err)
-
-		assert.NotNil(t, parsedToken)
-	})
+func init() {
+	AddSupportedAlgorithm(jwa.ES256K)
 }
