+++ crypto/jwx/jwx_es256k_test.go
@@ -18,16 +18,28 @@
  *
  */
 
-package oauth
+package jwx
 
 import (
+	"crypto"
+	"github.com/lestrrat-go/jwx/v2/jwa"
+	"github.com/lestrrat-go/jwx/v2/jwt"
+	"github.com/nuts-foundation/nuts-node/crypto/test"
 	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
 	"testing"
 )
 
 func TestES256k(t *testing.T) {
-	t.Run("supported formats specifies ES256K", func(t *testing.T) {
-		assert.Contains(t, DefaultOpenIDSupportedFormats()["jwt_vp_json"]["alg_values_supported"], "ES256K")
-		assert.Contains(t, DefaultOpenIDSupportedFormats()["jwt_vc_json"]["alg_values_supported"], "ES256K")
+	t.Run("test ES256K", func(t *testing.T) {
+		ecKey := test.GenerateECKey()
+		token := jwt.New()
+		signature, _ := jwt.Sign(token, jwt.WithKey(jwa.ES256K, ecKey))
+		parsedToken, err := ParseJWT(string(signature), func(_ string) (crypto.PublicKey, error) {
+			return ecKey.Public(), nil
+		})
+		require.NoError(t, err)
+
+		assert.NotNil(t, parsedToken)
 	})
 }
